+++ packages/flutter/lib/src/widgets/reorderable_list.dart
@@ -20,7 +20,7 @@ import 'scroll_view.dart';
 import 'scrollable.dart';
 import 'scrollable_helpers.dart';
 import 'sliver.dart';
-import 'sliver_explicit_extent_list.dart';
+import 'sliver_varied_extent_list.dart';
 import 'sliver_prototype_extent_list.dart';
 import 'ticker_provider.dart';
 import 'transitions.dart';
