+++ packages/flutter/lib/src/widgets/scroll_view.dart
@@ -23,7 +23,7 @@ import 'scroll_physics.dart';
 import 'scrollable.dart';
 import 'scrollable_helpers.dart';
 import 'sliver.dart';
-import 'sliver_explicit_extent_list.dart';
+import 'sliver_varied_extent_list.dart';
 import 'sliver_prototype_extent_list.dart';
 import 'viewport.dart';
