+++ packages/flutter/lib/widgets.dart
@@ -132,7 +132,7 @@ export 'src/widgets/shortcuts.dart';
 export 'src/widgets/single_child_scroll_view.dart';
 export 'src/widgets/size_changed_layout_notifier.dart';
 export 'src/widgets/sliver.dart';
-export 'src/widgets/sliver_explicit_extent_list.dart';
+export 'src/widgets/sliver_varied_extent_list.dart';
 export 'src/widgets/sliver_fill.dart';
 export 'src/widgets/sliver_layout_builder.dart';
 export 'src/widgets/sliver_persistent_header.dart';
